+++ src-tauri/crates/psysonic-audio/src/lib.rs
@@ -17,6 +17,7 @@ pub mod mix_commands;
 mod play_input;
 pub mod playback_rate;
 mod sink_swap;
+mod source_build;
 mod preserve_worker;
 pub mod preload_commands;
 pub(crate) mod progress_task;
